changes for mac compatibility
git-svn-id: https://vexx.googlecode.com/svn/trunk@814 e3525c50-fa1a-11de-8c8a-61c5e9addfc9

+++ alter2/plugins/UI/UIProfilerSurface.cpp
@@ -26,7 +26,7 @@ public:
     if(role == Qt::DisplayRole)
       {
       if(column == 0) return message();
-      if(column == 1) return statistics().count();
+      if(column == 1) return (quint64) statistics().count();
       if(column == 2) return statistics().total().milliseconds();
       if(column == 3) return statistics().minimum().milliseconds();
       if(column == 4) return statistics().average().milliseconds();
